DF-6053 add keyExist check for path_color_prop_name
Former-commit-id: 1ed6bb385cdfabd1337b92c171cee282dfcbfe87

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/PathSystem.cpp
@@ -104,7 +104,7 @@ void PathSystem::AddEntity(DAVA::Entity * entity)
     if (pc && pc->GetColor() == Color())
     {
         KeyedArchive *props = GetCustomPropertiesArchieve(entity);
-        if (props)
+        if (props && props->IsKeyExists(PATH_COLOR_PROP_NAME))
         {
             pc->SetColor(DAVA::Color(props->GetVector4(PATH_COLOR_PROP_NAME)));
             props->DeleteKey(PATH_COLOR_PROP_NAME);
